Adds support for changing the keyfile
This commit adds support for changing the keyfile used to connect to the
VMPooler boxes. Additionally, I updated the keyfile sync to only push up
the provided keyfile, and not the entire `.ssh` directory. Only problem
is, this makes it so that we cannot checkout github repos on the vm.